# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant PIM API
  version: 2.4.0
  description: |
    API para gestión de información de productos multi-tenant.

    ## Arquitectura Modular

    El sistema está organizado en módulos especializados:
    - **Health**: Estado del sistema
    - **Business Types**: Tipos de negocio
    - **Brands**: Gestión de marcas
    - **Products**: Gestión de productos y variantes
    - **Variants**: Gestión de variantes de productos (HITO 2.1)
    - **Categories**: Gestión de categorías
    - **Attributes**: Gestión de atributos
    - **Category-Attributes**: Relación categoría-atributo
    - **Global Catalog**: Catálogo global de productos con enrichment y quality score
    - **Global Catalog Public**: Endpoints públicos del catálogo global (sin auth)
    - **Product Requests**: Solicitudes de productos por tenants
    - **Quickstart**: Configuración rápida e importación desde global catalog
    - **Import**: Importación masiva con variantes (HITO 2.1)

    ## Actualizado v2.3.0

    - Global Catalog: CRUD completo, búsqueda por EAN, enrichment queue
    - Product Requests: Solicitudes de tenants con flujo approve/reject/fulfill
    - Quickstart: Import desde global catalog, import masivo por business type
    - Public endpoints: Búsqueda, sugerencias, facets sin autenticación
    - Variantes de productos: Product -> Variants con SKU único
    - Templates Quickstart: Configuración rápida por tipo de negocio
...
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticación
    tenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: ID del tenant para operaciones multi-tenant
    userRole:
      type: apiKey
      in: header
      name: X-User-Role
      description: |
        Rol del usuario para autorización:
        - `super_admin`: Acceso completo
        - `marketplace_admin`: Gestión de marketplace
        - `tenant_admin`: Administración del tenant
        - `tenant_user`: Usuario del tenant
...
paths:
...
  /business-types:
    get:
      summary: Listar tipos de negocio
      tags: [Business Types]
      description: Obtiene la lista de tipos de negocio disponibles
      security:
        - BearerAuth: []
        - tenantId: []
      parameters:
        - name: only_active
          in: query
          description: Solo tipos de negocio activos
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Lista de tipos de negocio
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    code:
                      type: string
                    name:
                      type: string
                    description:
                      type: string
                    icon:
                      type: string
                    color:
                      type: string
                    is_active:
                      type: boolean
                    sort_order:
                      type: integer
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor

    post:
      summary: Crear tipo de negocio
      tags: [Business Types]
      description: Crea un nuevo tipo de negocio (solo administradores)
      security:
        - BearerAuth: []
        - tenantId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                code:
                  type: string
                  example: "RESTAURANT"
                name:
                  type: string
                  example: "Restaurante"
                description:
                  type: string
                  example: "Tipo de negocio para restaurantes"
                icon:
                  type: string
                  example: "restaurant"
                color:
                  type: string
                  example: "#ff5722"
                sort_order:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Tipo de negocio creado exitosamente
        '400':
          description: Datos de entrada inválidos
        '401':
          description: No autorizado
        '409':
          description: Tipo de negocio ya existe
        '500':
          description: Error interno del servidor
  /business-types/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ID del tipo de negocio

    get:
      summary: Obtener tipo de negocio por ID
      tags: [Business Types]
      security:
        - BearerAuth: []
        - tenantId: []
      responses:
        '200':
          description: Detalles del tipo de negocio
        '404':
          description: Tipo de negocio no encontrado
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor

    put:
      summary: Actualizar tipo de negocio
      tags: [Business Types]
      security:
        - BearerAuth: []
        - tenantId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                icon:
                  type: string
                color:
                  type: string
                sort_order:
                  type: integer
      responses:
        '200':
          description: Tipo de negocio actualizado exitosamente
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Tipo de negocio no encontrado
        '500':
          description: Error interno del servidor

    delete:
      summary: Eliminar tipo de negocio
      tags: [Business Types]
      security:
        - BearerAuth: []
        - tenantId: []
      responses:
        '204':
          description: Tipo de negocio eliminado
        '404':
          description: Tipo de negocio no encontrado
        '409':
          description: No se puede eliminar (tiene datos asociados)
        '500':
          description: Error interno del servidor

  # ==========================================
  # BRANDS ENDPOINTS
  # ==========================================
...